requirements.txt
absl-py==1.0.0
asgiref==3.4.1
astunparse==1.6.3
blinker==1.4
boto3==1.19.12
botocore==1.22.12
cachetools==4.2.4
cchardet==2.1.7
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.7
click==8.0.3
cycler==0.11.0
Django==3.2.8
djangorestframework==3.12.4
dynamorm==0.11.0
fastprogress==1.0.0
filelock==3.3.2
flatbuffers==2.0
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googletrans==3.1.0a0
grpcio==1.41.1
h11==0.9.0
h2==3.2.0
h5py==3.5.0
hpack==3.0.0
hstspreload==2021.11.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.1.2
hyperframe==5.2.0
idna==2.10
jieba==0.42.1
jmespath==0.10.0
joblib==1.1.0
keras-bert==0.88.0
keras-embed-sim==0.9.0
keras-layer-normalization==0.15.0
keras-multi-head==0.28.0
keras-pos-embd==0.12.0
keras-position-wise-feed-forward==0.7.0
Keras-Preprocessing==1.1.2
keras-self-attention==0.50.0
keras-transformer==0.39.0
kiwisolver==1.3.2
langdetect==1.0.9
libclang==12.0.0
Markdown==3.3.4
marshmallow==3.14.0
matplotlib==3.4.3
numpy==1.21.4
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.2
pandas==1.3.4
Pillow==8.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.5
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
rfc3986==1.5.0
rsa==4.7.2
s3transfer==0.5.0
sacremoses==0.0.46
scikit-learn==0.23.2
scipy==1.7.2
sentencepiece==0.1.96
seqeval==0.0.19
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
syntok==1.3.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
threadpoolctl==3.0.0
tokenizers==0.10.3
tqdm==4.62.3
transformers==4.10.3
typing-extensions==3.10.0.2
urllib3==1.26.7
vaderSentiment==3.3.2
Werkzeug==2.0.2
Whoosh==2.7.4
wrapt==1.13.3
